Initial commit of current scripts:
SophosClientInstall_1.3.vbs - Sophos deployment script to automate deployment of Sophos from a sites EC install.
UIDConfig.xml - Sample configuration file to be used in conjunctio with UIDRADIUSscript to allow for ease of config changes
UIDRADIUSscript4.0.vbs - Script that parses RADIUS accounting logs on a site and submits relevant users to the Palo-Alto User-ID Agent
User-ID 802.1x Script.xml - Sample Exported task for the UID Script, allows for the Username to be passed from the Windows event.
logarchive.vbs - Quick and dirty script to allow for the archiving and deletion of old RADIUS accounting logs to prevent disk-space issues